Rename Appsignal.Phoenix.Plug to Appsignal.Plug
